refactor: move pages out of components; configure baseUrl
