+++ package.json
@@ -4,7 +4,6 @@
   "version": "3.0.0",
   "private": true,
   "author": "colinrobertbrooks@gmail.com",
-  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
@@ -38,7 +37,8 @@
     "cross-env": "^7.0.2",
     "npm-run-all": "^4.1.5",
     "typescript": "^5.0.2",
-    "vite": "^4.4.5"
+    "vite": "^4.4.5",
+    "vite-tsconfig-paths": "^4.0.7"
   },
   "engines": {
     "node": ">=18.17.0",
