+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './layout';
+export * from './styled';
+export * from './svg';
