+++ src/pages/Collection/consts.ts
