+++ src/styles/consts.ts
