+++ src/styles/utils.ts
@@ -1,3 +1,3 @@
-import { BASE_FONT_SIZE_PX } from './constants';
+import { BASE_FONT_SIZE_PX } from './consts';
 
 export const getRems = (px: number): string => `${px / BASE_FONT_SIZE_PX}rem`;
