+++ vite.config.ts
@@ -1,9 +1,10 @@
+import reactPlugin from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
-import react from '@vitejs/plugin-react';
+import tsconfigPathsPlugin from 'vite-tsconfig-paths';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [reactPlugin(), tsconfigPathsPlugin()],
   server: {
     port: 3000,
   },
